moved representation mapper to seperate file for easy plug and play for other developers
